refactor(evm): drop the json-rpc database from the public surface
